Conflicts:
	R-Type/server/sources/Network/Client.cpp

+++ R-Type/shared/sources/Commands/ICommand.cpp
@@ -1,4 +1,4 @@
 #include "ICommand.hpp"
 
-const int32_t				ICommand::MAGIC_CODE = 0x150407CA;
+const uint64_t		ICommand::MAGIC_CODE = 0x150407CA;
 const unsigned int	ICommand::HEADER_SIZE = sizeof(ICommand::Header);
