+++ R-Type/shared/sources/Commands/SharedCommandUpdateScore.cpp
@@ -6,7 +6,7 @@ CommandUpdateScore::CommandUpdateScore(void) : mId(0), mPseudo(""), mScore(0) {
 CommandUpdateScore::~CommandUpdateScore(void) {
 }
 
-int CommandUpdateScore::getId(void) const {
+uint64_t CommandUpdateScore::getId(void) const {
 	return mId;
 }
 
@@ -14,11 +14,11 @@ const std::string &CommandUpdateScore::getPseudo(void) const {
 	return mPseudo;
 }
 
-int CommandUpdateScore::getScore(void) const {
+uint64_t CommandUpdateScore::getScore(void) const {
 	return mScore;
 }
 
-void CommandUpdateScore::setId(int id) {
+void CommandUpdateScore::setId(uint64_t id) {
 	mId = id;
 }
 
@@ -26,7 +26,7 @@ void CommandUpdateScore::setPseudo(const std::string &pseudo) {
 	mPseudo = pseudo;
 }
 
-void CommandUpdateScore::setScore(int score) {
+void CommandUpdateScore::setScore(uint64_t score) {
 	mScore = score;
 }
